Fix all critical bugs: float/double mismatch, DataInputStream reuse, async result wait, CnxBD Singleton, duplicate close, build classpath, docs
Agent-Logs-Url: https://github.com/sutterread/CombateDeSumos/sessions/f952f75d-b838-4268-a280-fe47a35d8982

+++ src/pa/combatedesumos/Servidor/Control/ControlLuchador.java
@@ -72,7 +72,7 @@ public class ControlLuchador {
      * @param peso      peso del luchador
      * @param kimarites técnicas del luchador
      */
-    public void insertarLuchadorEnBD(String nombre, double peso, String[] kimarites) {
+    public void insertarLuchadorEnBD(String nombre, float peso, String[] kimarites) {
         try {
             LuchadorDTO luchador = new LuchadorDTO(nombre, peso, 0, kimarites);
             ldao.insertarLuchador(luchador);
